backend/pyproject.toml
[project]
name = "sigdoc"
version = "0.1.0"
description = "Document template management and generation system"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.1",
    "docxtpl>=0.18.0",
    "python-docx<1.1.1",
    "openpyxl>=3.1.0",
    "minio>=7.2.0",
    "python-multipart>=0.0.9",
    "slowapi>=0.1.9",
    "cachetools>=5.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src", "."]
